release cast

+++ philo_two/src/philosopher2.c
@@ -6,7 +6,7 @@
 /*   By: sgah <sgah@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/26 19:03:23 by sgah              #+#    #+#             */
-/*   Updated: 2021/04/27 13:18:20 by sgah             ###   ########.fr       */
+/*   Updated: 2021/04/27 14:43:42 by sgah             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ t_philo
 	*change_state(t_philo *philo, int state, t_philosopher *info)
 {
 	philo->state = state;
-	if (philo->info->is_dead != 0 || info->limit_eat == (int)info->nb_philo)
+	if (philo->info->is_dead != 0 || info->limit_eat == info->nb_philo)
 		return (philo);
 	printf("%u %i %s\n", time_lapse(philo->info->start),
 	philo->id, g_msg[state]);
@@ -29,9 +29,9 @@ int
 	if (info->limit_eat == -1)
 		return (0);
 	philo->eat++;
-	if (philo->eat == (int)info->nb_eat)
+	if (philo->eat == info->nb_eat)
 		info->limit_eat++;
-	if (info->limit_eat == (int)info->nb_philo)
+	if (info->limit_eat == info->nb_philo)
 		return (1);
 	return (0);
 }
@@ -39,7 +39,7 @@ int
 t_philo
 	*routine(t_philo *philo, t_philosopher *info)
 {
-	if (info->is_dead != 0 || info->limit_eat == (int)info->nb_philo)
+	if (info->is_dead != 0 || info->limit_eat == info->nb_philo)
 		return (philo);
 	take_fork(philo, info);
 	philo = change_state(philo, EAT, info);
@@ -47,12 +47,12 @@ t_philo
 		return (free_fork(philo, info));
 	philo->last_eat = time_lapse(info->start);
 	ft_usleep(info->time_eat);
-	if (info->is_dead != 0 || info->limit_eat == (int)info->nb_philo)
+	if (info->is_dead != 0 || info->limit_eat == info->nb_philo)
 		return (free_fork(philo, info));
 	philo = free_fork(philo, info);
 	philo = change_state(philo, SLEEP, info);
 	ft_usleep(info->time_sleep);
-	if (info->is_dead != 0 || info->limit_eat == (int)info->nb_philo)
+	if (info->is_dead != 0 || info->limit_eat == info->nb_philo)
 		return (philo);
 	philo = change_state(philo, THINK, info);
 	return (philo);
@@ -70,7 +70,7 @@ void
 		;
 	if (philo->id % 2 == 0)
 		usleep(1000);
-	while (info->is_dead == 0 && info->limit_eat != (int)info->nb_philo)
+	while (info->is_dead == 0 && info->limit_eat != info->nb_philo)
 		philo = routine(philo, info);
 	return (data);
 }
